Update typedoc.json
Setting the mode to "file" creates a much more readable HTML layout for Typedoc, for an example see https://skyra-project.github.io/skyra-decorators/, comparing it to for example https://dirigeants.github.io/klasa/. The difference being of course that with "modules" you get an extremely long list of files, most of which only have 1 or 2 exports. Furthermore the layout of "files" is much more in line with the layout of a page such as the "Classes" tab on https://klasa.js.org/#/docs/klasa/settings/Getting%20Started/GettingStarted.